feat: add prisma, supabase, and tailwind setup #3
