+++ next.config.ts
@@ -1,7 +1,8 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  // Next.js will automatically use the src directory if it exists
+  // No explicit config needed for src/ structure in Next.js 13+
 };
 
 export default nextConfig;
