+++ src/app/page.tsx
@@ -1,0 +1,8 @@
+export default function Home() {
+  return (
+    <div className="text-3xk font-bold text-center">
+      Home
+      <p className="text-primary">Dark Mode</p>
+    </div>
+  );
+}
